Add Task Lists CRUD

+++ app/Http/Controllers/TaskListController.php
@@ -10,7 +10,7 @@ class TaskListController extends Controller
 {
     public function index()
     {
-        return TaskListResource::collection(TaskList::all());
+        return TaskListResource::collection(TaskList::where('user_id', auth()->id())->get());
     }
 
     public function store(TaskListRequest $request)
@@ -20,18 +20,23 @@ class TaskListController extends Controller
 
     public function show(TaskList $taskList)
     {
+        if ($taskList->user_id !== auth()->id()) abort(403);
+
         return new TaskListResource($taskList);
     }
 
     public function update(TaskListRequest $request, TaskList $taskList)
     {
-        $taskList->update($request->validated());
+        if ($taskList->user_id !== auth()->id()) abort(403);
 
+        $taskList->update($request->validated());
         return new TaskListResource($taskList);
     }
 
     public function destroy(TaskList $taskList)
     {
+        if ($taskList->user_id !== auth()->id()) abort(403);
+
         $taskList->delete();
 
         return response()->json();
